core/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
boto3==1.34.60
botocore==1.34.60
certifi==2024.2.2
cffi==1.16.0
decorator==5.1.1
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-storages==1.14.2
django-versatileimagefield==3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
exceptiongroup==1.2.0
executing==2.0.1
gunicorn==21.2.0
inflection==0.5.1
ipython==8.22.2
jedi==0.19.1
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.6
minio==7.2.5
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
s3transfer==0.10.0
sendgrid==6.11.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
starkbank-ecdsa==2.2.0
traitlets==5.14.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.13
